# weathra/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CallRole": {
        "description": "The role a language model call serves, which is what a policy maps against.\n\nThe role rather than the node (design.md decision 22): a fourth node needing a structured\ndecision reuses ``ROUTING`` and no policy changes. ``LAB`` is separate because a comparison run\nis internal traffic accounted against the internal allowance, never a product plan's.",
        "enum": [
          "routing",
          "synthesis",
          "lab"
        ],
        "title": "CallRole",
        "type": "string"
      },
...
      "CatalogListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "count",
          "entries"
        ],
        "title": "CatalogListResponse",
        "type": "object"
      },
      "CatalogStatus": {
        "description": "Whether an entry may be resolved at all.\n\n``specs/model-catalog`` treats a disabled entry as unavailable everywhere \u2014 no policy resolves\nit, no override accepts it, no lab selection offers it \u2014 while its recorded history stays\nreadable and attributed. Disabling is therefore the sanctioned response to a model that has\nstopped working, and is why nothing in this package deletes a catalog row.",
        "enum": [
          "enabled",
          "disabled"
        ],
        "title": "CatalogStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanCode": {
        "description": "The canonical subscription plan codes.\n\nThree tiers, lowercase, stable enough to map to an external billing product later\n(``specs/usage-limits``' \"room left for a later billing integration\"). The display names live\nin ``subscription_plans`` as data, so renaming *Premium* on a pricing page is a row update; the\ncode here is the identifier that must not move, because `user_plans` rows and recorded usage\nevents both carry it.\n\nThere is deliberately **no `plus`**. An earlier draft of the specs named the middle tier that\nway; the product decision of 2026-09-09 settled on Free / Pro / Premium and the specs were\nreconciled to match. No alias is provided in either direction \u2014 this foundation has never\nshipped, so there is no customer contract to be backward-compatible with, and an alias would\nonly make two names for one tier that a later reader has to reconcile again.",
        "enum": [
          "free",
          "pro",
          "premium"
        ],
        "title": "PlanCode",
        "type": "string"
      },
...
      "PlanMappingRequest": {
        "additionalProperties": false,
        "description": "Which policy a plan resolves to, per call role.",
        "properties": {
          "policy_by_call_role": {
            "additionalProperties": {
              "type": "string"
            },
            "minProperties": 1,
            "propertyNames": {
              "$ref": "#/components/schemas/CallRole"
            },
            "title": "Policy By Call Role",
            "type": "object"
          }
        },
        "required": [
          "policy_by_call_role"
        ],
        "title": "PlanMappingRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SupabaseAccessToken": {
        "bearerFormat": "JWT",
        "description": "A Supabase Auth access token, presented as `Authorization: Bearer <token>`. Weathra validates the signature against the project's published signing keys and never issues, stores, or refreshes a credential of its own.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agentic weather intelligence. Every figure in a Weathra answer is retrieved from a named provider\nor computed by deterministic code, and every answer carries the evidence record to prove it.\n\nEndpoints are classified **public** or **protected**. A public endpoint takes parameters and\nreturns weather; it reads no user-owned data, except that it applies a signed-in caller's unit\npreference. A protected endpoint requires a Supabase access token as a Bearer credential and\noperates on the acting user's own data.\n",
    "title": "Weathra",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/models": {
      "get": {
        "description": "**Protected. Administrative.** Lists and creates model catalog entries.\n\nEvery catalog entry, optionally narrowed by status and by capability role.\n\nIncludes disabled entries, which is the point of the status filter: `specs/model-catalog`\nrequires a withdrawn model to stay auditable rather than disappearing, so an administrator can\nstill see what served last month.",
        "operationId": "list_catalog_api_v1_admin_models_get",
        "parameters": [
          {
            "description": "Narrow by status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CatalogStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow by status.",
              "title": "Status"
            }
          },
          {
            "description": "Narrow to entries fit for one call role.",
            "in": "query",
            "name": "capability_role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CallRole"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow to entries fit for one call role.",
              "title": "Capability Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "List the model catalog",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "description": "**Protected. Administrative.** Lists and creates model catalog entries.",
        "operationId": "create_catalog_entry_api_v1_admin_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Add a model to the catalog",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/models/{catalog_key}": {
      "patch": {
        "description": "**Protected. Administrative.** Edits one model catalog entry.",
        "operationId": "edit_catalog_entry_api_v1_admin_models__catalog_key__patch",
        "parameters": [
          {
            "description": "The catalog entry.",
            "in": "path",
            "name": "catalog_key",
            "required": true,
            "schema": {
              "description": "The catalog entry.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Catalog Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Edit a catalog entry",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/models/{catalog_key}/disable": {
      "post": {
        "description": "**Protected. Administrative.** Withdraws a model from resolution, refused for the last one serving a call role.\n\nWithdraw a model from resolution.\n\nRefused when it is the last enabled entry serving a call role: `specs/model-catalog` requires\nthe system to keep answering, and a disable that left routing with nothing to resolve would\ntake the agent surface down from an administrative screen.",
        "operationId": "disable_catalog_entry_api_v1_admin_models__catalog_key__disable_post",
        "parameters": [
          {
            "description": "The catalog entry.",
            "in": "path",
            "name": "catalog_key",
            "required": true,
            "schema": {
              "description": "The catalog entry.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Catalog Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Disable a catalog entry",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/models/{catalog_key}/enable": {
      "post": {
        "description": "**Protected. Administrative.** Returns a model to resolution.",
        "operationId": "enable_catalog_entry_api_v1_admin_models__catalog_key__enable_post",
        "parameters": [
          {
            "description": "The catalog entry.",
            "in": "path",
            "name": "catalog_key",
            "required": true,
            "schema": {
              "description": "The catalog entry.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Catalog Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Enable a catalog entry",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/admin/plans/{plan_code}/policies": {
      "put": {
        "description": "**Protected. Administrative.** Re-points a plan at different policies, per call role.\n\nWhich policy each of a plan's call roles resolves to.\n\nThe plan code is a ``PlanCode``, so a request naming ``plus`` is refused by validation before\nthe handler runs: the retired tier is not a value this enum has, and no alias admits it.",
        "operationId": "set_plan_policy_mapping_api_v1_admin_plans__plan_code__policies_put",
        "parameters": [
          {
            "description": "free, pro, or premium.",
            "in": "path",
            "name": "plan_code",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlanCode",
              "description": "free, pro, or premium."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanMappingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Re-point a plan at different policies",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/policies": {
      "get": {
        "description": "**Protected. Administrative.** Lists and creates model policies.",
        "operationId": "list_policies_api_v1_admin_policies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "List the model policies",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "description": "**Protected. Administrative.** Lists and creates model policies.",
        "operationId": "create_policy_api_v1_admin_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Create a model policy",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/policies/{policy_id}/candidates": {
      "put": {
        "description": "**Protected. Administrative.** Re-points a policy's ordered candidate list \u2014 a model promotion.",
        "operationId": "set_policy_candidates_api_v1_admin_policies__policy_id__candidates_put",
        "parameters": [
          {
            "description": "The policy.",
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "description": "The policy.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Policy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCandidatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Re-point a policy's candidate list",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/policies/{policy_id}/fallback": {
      "put": {
        "description": "**Protected. Administrative.** Sets or clears a policy's declared fallback.",
        "operationId": "set_policy_fallback_api_v1_admin_policies__policy_id__fallback_put",
        "parameters": [
          {
            "description": "The policy.",
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "description": "The policy.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Policy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyFallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Set or clear a policy's fallback",
        "tags": [
          "administration"
        ]
      }
    },
...
  }
}